requirements.txt
django==3.2.7
gunicorn
pillow
django-environ
django-filter
drf_yasg
djangorestframework
django-cors-headers
psycopg2-binary
celery
django-redis
djangorestframework-simplejwt
djoser